cleanse and refactor lora testing suite.
